택시 기하학(수정)
https://www.acmicpc.net/problem/3053

+++ N3053.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class N3053 {
     public static void main(String[]args){
         Scanner scan=new Scanner(System.in);
-        double n= scan.nextInt();
+        double n= scan.nextDouble();
         System.out.println(String.format("%.6f",Math.pow(n,2)*Math.PI));
         System.out.println(String.format("%.6f",Math.pow(n,2)*2));
     }
